+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Tweak user roles (currently can have only one role)
 - Create timetable for animal walks
 - Animal page contains calendar with walks
+- Open walk by clicking in calendar and register it